JBIDE-1406 - allowing version id's for the descriptor and associated UI changes and test
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@9584 a97e2381-89e5-4abb-bab3-167db6db766c

+++ core/plugins/org.jboss.ide.eclipse.archives.core/src/main/org/jboss/ide/eclipse/archives/core/model/internal/xb/XbPackages.java
@@ -25,7 +25,7 @@ import java.util.List;
 
 public class XbPackages extends XbPackageNodeWithProperties {
 	
-	private float version;
+	private double version;
 	public XbPackages () {
 		super();
 	}
@@ -47,11 +47,11 @@ public class XbPackages extends XbPackageNodeWithProperties {
 		return getChildren(XbPackage.class);
 	}
 
-	public float getVersion() {
+	public double getVersion() {
 		return version;
 	}
 
-	public void setVersion(float version) {
+	public void setVersion(double version) {
 		this.version = version;
 	}
 }
